requirements.txt
asgiref==3.3.1
boto3==1.17.22
Django==3.1.5
django-cors-headers==3.7.0
django-smtp-ssl==1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
paypal-checkout-serversdk==1.0.1
psycopg2-binary==2.8.6
pyjwt==2.0.1
python-decouple==3.4
pytz==2020.5
shapely==1.7.1
sqlparse==0.4.1
swagger-spec-validator==2.7.3
